+++ src/Ci/Codeship.php
@@ -33,12 +33,12 @@ class Codeship extends AbstractCi
         return $this->env->getString('CI_BUILD_URL');
     }
 
-    public function getGitCommit(): string
+    public function getCommit(): string
     {
         return $this->env->getString('COMMIT_ID');
     }
 
-    public function getGitBranch(): string
+    public function getBranch(): string
     {
         return $this->env->getString('CI_BRANCH');
     }