+++ src/Ci/Jenkins.php
@@ -33,12 +33,12 @@ class Jenkins extends AbstractCi
         return $this->env->getString('BUILD_URL');
     }
 
-    public function getGitCommit(): string
+    public function getCommit(): string
     {
         return $this->env->getString('GIT_COMMIT');
     }
 
-    public function getGitBranch(): string
+    public function getBranch(): string
     {
         return $this->env->getString('GIT_BRANCH');
     }